# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /meta/status/{clientId}:
    get:
      operationId: getMetaStatus
      tags: [meta]
      summary: Get Meta connection status for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meta status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaStatus"
  /meta/campaigns/{clientId}:
    get:
      operationId: getMetaCampaigns
      tags: [meta]
      summary: Fetch campaigns from Meta Ads API
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meta campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCampaignsResponse"
components:
  schemas:
...
    Notification:
      type: object
      required: [id, title, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [alert, info, warning, success]
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string

    # Slack
...
    MetaStatus:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
        accountId:
          type: string
        scopes:
          type: string
        expiresAt:
          type: ["string", "null"]
    MetaCampaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        objective:
          type: string
        daily_budget:
          type: string
        spend:
          type: string
    MetaCampaignsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MetaCampaign"
        error:
          type: object
          properties:
            message:
              type: string